[svn r11] Add C, gamma, and degree options to SVMs Add XML writer Add mostly working OntoShell server code Bugfix typing and inheritance in property files Improve OntoShell server/console inheritance Remove unnecessary buggy hack code from XML vacuum

+++ logrus/libBioUtils/mathb.h
@@ -7,7 +7,7 @@ namespace libBioUtils {
 
 class CMath : CMathImpl {
 public:
-	static unsigned int GCD( unsigned int, unsigned int );
+	static size_t GCD( size_t, size_t );
 	static double Min( double, double );
 	static size_t Round( double );
 	static double LogFact( size_t );
